SERVER-25717 Allow for negative $push indexes.
It is now possible to issue $push updates with negative $position
indexes. These will be treated as an offset from the end of the array.

For example, given the following document:

{acl: [
    {grant: true, user: "amcgregor"},
    {grant: true, user: "bdole"},
    {grant: false}
]}

Issuing the following update:

{$push: {acl: {$each: [{grant: true, user: "algore"}], $position: -1}}}

Will result in:

{acl: [
    {grant: true, user: "amcgregor"},
    {grant: true, user: "bdole"},
    {grant: true, user: "algore"},
    {grant: false}
]}

This matches Python list manipulation semantics.

+++ src/mongo/db/ops/modifier_push.h
@@ -110,7 +110,7 @@ private:
     bool _slicePresent;
     int64_t _slice;
     bool _sortPresent;
-    size_t _startPosition;
+    int _startPosition;
 
     PatternElementCmp _sort;
